config: clarify Linux mount options
The former wording was inconsistent with runc, crun, youki, etc.

Several `mount(8)` options such as `X-mount.mode` are unimplemented in those OCI runtimes,
and vice versa (`tmpcopyup`, `rro`, etc.)

The option table added in this commit conforms to runc 1.1, but `acl`
and `noacl` are excluded as the runc implementation seems wrong (runc issue 3738).

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>