config-linux: clarify the "MB:"-line filtering in RDT.
The thinking is that the runtimes should not do the filtering of values,
but instead just apply the values in order. This way the possible
MB-lines in l3CacheSchema will become overwritten by memBwSchema values
(if the domains overlap).

Note that we can't just concatenate the values because kernel will error
out if the same domain is attempted to be set multiple times within one
write() call.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>