config: add support for org.opencontainers.image annotations
These annotations are specified as part of the official conversion
process from an OCI image configuration to an OCI runtime configuration
since v1.0.0-rc7 of the image-spec but they were never officially
specified in the runtime-spec.

The fact these are not allowed by the current runtime-spec causes some
issues with inter-spec compatibility. In order for tools like umoci to
be able to do this generation properly, we need to allow this namespace
to be used.

Ref: https://github.com/opencontainers/image-spec/blob/v1.1.0-rc2/conversion.md
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>