config: Add Hardware description object to the VM configuration
This adds section to describe HW that should be passed through to the
VM. This enables Hardware-level isolation provided by XEN for
e.g. functional safety use cases.

Adds hwConfig object to the VM section which is apt to describe the
initial configuration for the VM, sush as number of vcpus and memory,
provided to the VM.
Hardware description includes path to the device-tree, that should be
passed to the VM and the hardware configuration parameters which
provides all needed data for VM to use the devices, such as:
  - dtdevs: host device tree nodes to passthrough to the VM;
  - iomems: allow auto-translated domains to access specific hardware
  I/O memory pages;
  - irqs: allows VM to access specific physical IRQs.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>