mount: Allow relative mount destinations on Linux
We tried to make runc enforce abs dest path several times, and always
had to revert it due to some tools not yet doing it. The last occurrence
is this one:
	https://github.com/opencontainers/runc/issues/3944#issuecomment-1669261383

I don't see any reason to force abs dst paths on Linux, as far as I know
there is no security bug nor anything. Let's just relax the spec
wording, matching all the runtimes behavior when the paths is relative,
and be done with it.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>