config: improve bind mount and propagation doc
Current documentation talks about bind/rbind and various propagation
flags (shared, slave, private, unbindable etc.) as being filesystem-
specific options, and refers to mount(8) section describing
filesystem-specific options.

That is wrong, because bind mounts are a separate class of mounts,
not related to file systems.

Fix the confusion, and link to sharedsubtree.txt kernel document
as it happens to be the best source of information about the topic
(not counting the kernel sources).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>