specs-go: add missing deprecation comment for Hooks.Prestart
Hooks.Prestart was deprecated in c1662686cff159595277b79322d0272f5182941b
(v1.0.2), but the field did not have a correctly formatted deprecation comment.

This patch updates the field's GoDoc to have a correctly formatted deprecation
comment, which will allow linters and IDEs to detect the deprecation status
of this feature.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>