specs-go: mark LinuxMemory.Kernel as deprecated
Support for kernel-memory limits was deprecated in the kernel, and documented
as "NOT RECOMMENDED" (or "SHOULD NOT" use) in  v1.1.0-rc.1 through commit
f02cd4a4273f5cf4898df4feee7dcf0d08756029.

This patch marks the field as deprecated in the go implementation of the
spec, so that linters and editors produces a warning and consumers get
notified of its status.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>