config: simplify final CPU affinity rule
Description of execCPUAffinity.final said that if it's not set or empty,
the final affinity is the one of container's cgroup. This was done
because we thought the kernel changes process' CPU affinity to one of
cgroup. It's not the case.

To keep the runtime simple, it makes sense to not do anything if the
final affinity is not explicitly set. This change does just that.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>