Add minimum supported Go version to CI (#1303)
* Add minimum supported Go version to CI

On top of automatically testing against the two most recent releases (what Go upstream supports), also test explicitly against our lower bound.

As noted in the previous change, don't have a `go.mod` to source this information from, so it's simply hard-coded in this file instead.

(I chose 1.21 as that was the lowest version we were testing against previously, but it's possible that could go lower or actually reasonably needs to go higher.)

Signed-off-by: Tianon Gravi <admwiggin@gmail.com>

* Add explicit `GOTOOLCHAIN=local` in CI

Signed-off-by: Tianon Gravi <admwiggin@gmail.com>

---------

Signed-off-by: Tianon Gravi <admwiggin@gmail.com>